# usuariosflask/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: API de Netflux realizada por el grupo 3 de la asignatura ASEE. Microservicio
    de **Usuarios**.
  license:
    name: Copyleft
    url: https://url/a/la/licencia
  title: Netflux Grupo 3 - OpenAPI 3.0
  version: 0.0.1
servers:
- url: https://path/api/v1
...
paths:
...
  /usuarios/{userId}/perfiles:
    get:
      description: Lista los perfiles de un usuario
      operationId: obtener_perfiles
      parameters:
      - description: ID del usuario a eliminar
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Perfil'
                type: array
          description: Perfiles obtenidos exitosamente
        "404":
          description: Usuario no encontrado
        "500":
          description: Error interno del servidor
      summary: Obtiene todos los perfiles del usuario especificado
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
    post:
      description: Crea un nuevo perfil para el usuario
      operationId: crear_perfil
      parameters:
      - description: ID del usuario a eliminar
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
        description: Objeto del perfil a crear
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Perfil creado exitosamente
        "500":
          description: Error interno del servidor
      summary: Añade un nuevo perfil al usuario especificado
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
  /usuarios/{userId}/perfiles/{profileId}:
    delete:
      description: Borra el perfil especificado de un usuario
      operationId: borrar_perfil_usuario
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil a obtener
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Perfil borrado exitosamente
        "403":
          description: El cliente no tiene permisos suficientes
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error interno del servidor
      summary: Borra el perfil especificado
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
    get:
      description: Obtiene el perfil específico de un usuario concreto
      operationId: obtener_perfil_usuario
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil a obtener
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Perfil obtenido exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error interno del servidor
      summary: Obtiene el perfil específico de un usuario concreto
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
    put:
      description: Actualiza el perfil especificado de un usuario
      operationId: actualizar_perfil_usuario
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil a obtener
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerfilUpdate'
        description: Objeto del perfil con la información actualizada
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Perfil actualizado exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error interno del servidor
      summary: Actualiza el perfil especificado
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
  /usuarios/{userId}/perfiles/{profileId}/historial:
    delete:
      description: Borra el historial de un perfil
      operationId: eliminar_contenido_historial
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil específico
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Historial borrado exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error del servidor
      summary: Borra el historial de un perfil
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
    get:
      description: Obtiene el historial de contenido completado por un perfil. Esta
        lista contendrá las series o películas terminadas de ver por el perfil.
      operationId: obtener_historial_perfil
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil específico
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Serie'
                type: array
          description: Historial obtenido exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error del servidor
      summary: Obtiene el historial de contenido completado por de un perfil
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
    post:
      description: Añade un nuevo contenido al historial de un perfil
      operationId: agregar_contenido_historial
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil específico
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Capitulo'
        description: Objeto del contenido a añadir
        required: true
      responses:
        "201":
          description: Contenido añadido exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error del servidor
      summary: Añade un nuevo contenido al historial de un perfil
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
  /usuarios/{userId}/perfiles/{profileId}/historial/{contentId}:
    patch:
      description: Actualiza la fecha de visualización de un contenido del historial de un perfil
      operationId: actualizar_contenido_historial
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil específico
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del contenido específico
        explode: false
        in: path
        name: contentId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Capitulo'
        description: Objeto del contenido a actualizar
        required: true
      responses:
        "200":
          description: Contenido actualizado exitosamente
        "404":
          description: Usuario, perfil o contenido no encontrado
        "500":
          description: Error del servidor
      summary: Actualiza un contenido del historial de un perfil
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
  /usuarios/{userId}/perfiles/{profileId}/lista:
    get:
      description: Obtiene la lista de contenidos guardados para ver de un perfil
      operationId: obtener_lista_perfil
      parameters:
      - description: ID del usuario específicado
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      - description: ID del perfil específico
        explode: false
        in: path
        name: profileId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Serie'
                type: array
          description: Lista obtenida exitosamente
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error del servidor
      summary: Obtiene la lista de un perfil concreto
      tags:
      - Perfil
      x-openapi-router-controller: openapi_server.controllers.perfil_controller
...
components:
  schemas:
...
    PerfilUpdate:
      allOf:
      - $ref: '#/components/schemas/Perfil'
      - required:
        - id
        - user_id
      example:
        user_id: 3
        preferencias_contenido:
          languages:
          - Espanol
          genres:
          - Action
          - Supernatural
        foto_perfil: path/to/image
        id: 1
        nombre: Euler
        historial_vistos:
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        mi_lista:
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
      title: PerfilUpdate
...